requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
openai-whisper>=20230314
transformers>=4.30.0
gradio==4.44.0
numpy>=1.24.0
scikit-learn>=1.3.0
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.11.0
requests>=2.31.0
pandas>=2.0.0
matplotlib>=3.7.0

